Linux: Check maximum send size up front
Only try sending the entire message in one packet if it will actually
fit.

This saves a syscall and some processing, but only for large
(fragmented) messages -- so it doesn't have a noticable performance
impact. However, it should make behaviour clearer and more predictable;
and it is also required in order to enable further cleanups.