Linux: Cleanup: Drop unnecessary intermediate casts
While it might have been necessary at some point in the cast between
`*mut` and `*const` in a separate step from casting between different
target types, it is definitely not needed with current Rust.

One might argue that doing it in a separate step makes the change more
explicit. However, in some cases it's obvious anyway (when the source is
`.as_ptr()` or `.as_mut_prt()`) -- and even where it's not quite as
obvious, it's not clear at all that the unnecessary extra step isn't
actually adding more confusion...