Add memfd feature to use memfd_create for linux
For linux OSes allow the use of memfd_create for the adventurous souls
that dare to do so. This feature should be enabled with the memfd
features flag and should only be available on linux.

Note: glibc does not currently provide a wrapper for this syscall yet.
As a result, it is recommended to use `syscall` directly. See the man
pages for details.